Add files via upload
File structure of project
app.py and wtf.py are in the same directory level as a directory called static and a directory called templates.
Inside of the templates directory is all of the .html files.
Inside of the static directory is a second directory called pictures. Inside of the pictures directory is all the logo .jpg and .png files.