Add files via upload
James Keifer Hudson
Changes made to each file:
wtf.py- Added the ability for the user to specify the endyear, endmonth, and enddate of their stock market lookup to the DateTicke wt form. 
Company_Page.html- changed it to be able to show the changes made to DateTicker form and display these changes directly. Also added a "message" section in order to display error messages to the user if they input an invalid date.
app.py- implemented changed that were made to DateTicker form. Also added lots of error checking to the data being submitted to the DateTicker form. This error checking utilizes a new dictionary called monthdaydict that stores the number of days in each month and a new function (made by Andres and diego) that compares two dates to see which one comes first. This is used to make sure that both the start date and end date come before the current day (or are the current day) and to ensure that the start date comes before the end date. Also made it so that the end date section of the DateTicker form is prefilled in with the current day for ease of use for the user. The user can of course delete this and specify their own end date if they choose too.